requirements.txt
numpy>=1.23.2
pandas>=1.0.0
scikit-learn>=1.0.0
matplotlib>=3.0.0
seaborn>=0.12.0
xgboost>=1.7.0
ucimlrepo>=0.0.7
joblib>=1.0.0
jupyter>=1.0.0

# API服务
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# 实验跟踪和模型管理
wandb>=0.12.0
mlflow>=1.20.0

# 数据处理
python-dotenv>=0.19.0
pyyaml>=5.4.1

# 模型序列化
pickle5>=0.0.11; python_version < '3.8' 